+++ Tavern/SwipingFunctionality/SwipingPage.xaml.cs
@@ -1,11 +1,14 @@
 
 
+using CommunityToolkit.Maui.Views;
+
 namespace Tavern.SwipingFunctionality;
 
 public partial class SwipingPage : ContentPage
 {
 	private Queue<Group> groups = new Queue<Group>(10);
     string connectionString = null;
+	
 
 	public SwipingPage()
 	{
@@ -30,33 +33,7 @@ public partial class SwipingPage : ContentPage
 			Navigation.PopAsync();
 		}	
 	}
-    private async void SwipedRight(Group likedGroup)
-    {
-        await LikeGroup(likedGroup);
-    }
-
-    private async Task LikeGroup(Group likedGroup)
-    {
-        ProfileSingleton singleton = ProfileSingleton.GetInstance();
-
-		using (var client = new HttpClient())
-		{
-			var parameters = new Dictionary<string, string>
-			{
-				{ "groupId", likedGroup.GroupId.ToString() },
-				{ "userId", singleton.ProfileId.ToString() }
-			};
-			var content = new FormUrlEncodedContent(parameters);
-			var response = await client.PostAsync(connectionString, content);
-			if (response.IsSuccessStatusCode)
-			{
-				ShowNextGroup();
-			}
-		}
-			//Group.JoinRequest(likedGroup.GroupId, singleton.ProfileId);
-
-        // Show the next group after liking
-    }
+    
 
     /**
 	 * PopulateGroupQueue - Helper function that populates the queue
