+++ Tavern/SwipingFunctionality/SwipingView.xaml
@@ -7,7 +7,7 @@
             x:Name="swipeLeft"
             Text="Like"
             BackgroundColor="Green"
-            Invoked="OnJoinGroup"
+            Invoked="swipeRight"
             />
     </SwipeView.LeftItems>
     <SwipeView.RightItems>
